fix(B-301): rename training feedback files to session feedback
